Stricter checks for mg_url_*()

+++ src/util.c
@@ -332,6 +332,6 @@ unsigned long mg_millis(void) {
 #else
   struct timespec ts;
   clock_gettime(CLOCK_REALTIME, &ts);
-  return ts.tv_sec * 1000 + ts.tv_nsec / 1000000;
+  return (unsigned long) ((uint64_t) ts.tv_sec * 1000 + ts.tv_nsec / 1000000);
 #endif
 }
